Split template types into files
