+++ aardwolf-templates/src/shortcuts.rs
@@ -1,0 +1,7 @@
+use gettext::Catalog;
+
+pub struct Shortcuts<'a> {
+    pub(crate) catalog: &'a Catalog,
+    pub(crate) profile_link: &'a str,
+    pub(crate) username: &'a str,
+}
